Rename brandAddresses directory
